requirements.txt
django
djangorestframework
requests
pylint[spelling]